+++ packages/metadata/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the metadata will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v1.6.4]
+### Added
+- Arc Testnet deployment
+
 ### Changed
 - Removed discontinued networks: scroll-sepolia, degenchain